refactor: split the library API out of admin
/api/library_stats and /api/sync_library count songs and start a scan.
They sat in admin because the info page's dashboard calls them, not
because they belong beside the routes that halt, reboot and
authenticate.

Moving them leaves admin with no /api/ route, so it goes internal and
its eight pages and redirects leave the spec: 36 documented paths down
to 28, and the non-/api remainder from 13 to 5.

URLs and view names are unchanged; the three url_for call sites in
info.html follow the endpoint. test_admin_routes probes
/api/library_stats for its 403, so its app registers the new blueprint
rather than swapping in a route that reboots the machine.
