+++ pikaraoke/templates/info.html
@@ -257,7 +257,7 @@
 			});
 
 			// Song library stats
-			$.getJSON("{{ url_for('admin.library_stats') }}", function(data) {
+			$.getJSON("{{ url_for('library.library_stats') }}", function(data) {
 				if (!data.error) {
 					$("#library-song-count").text(data.song_count);
 				}
@@ -267,7 +267,7 @@
 			$("#sync-library-btn").click(function() {
 				var $btn = $(this);
 				$btn.addClass("is-loading").prop("disabled", true);
-				$.post("{{ url_for('admin.sync_library') }}", function(data) {
+				$.post("{{ url_for('library.sync_library') }}", function(data) {
 					if (data.status !== "already_syncing") return;
 					// stays disabled -- sync_finished will re-enable
 				}).fail(function() {
@@ -297,7 +297,7 @@
 			});
 			window.socket.on("sync_finished", function() {
 				$("#sync-library-btn").removeClass("is-loading").prop("disabled", false);
-				$.getJSON("{{ url_for('admin.library_stats') }}", function(data) {
+				$.getJSON("{{ url_for('library.library_stats') }}", function(data) {
 					if (!data.error) $("#library-song-count").text(data.song_count);
 				});
 				showNotification({{ _('Library sync complete') | tojson }}, "is-success");
